backend/requirements.txt
# Core Django & Database
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.0
PyMySQL==1.1.0
django-environ==0.11.2
django-allauth==0.57.0
cryptography==42.0.4

# Async Tasks (Celery)
celery==5.3.6
redis==5.0.1

# AI Integration
requests==2.31.0
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2

# Web Scraping
beautifulsoup4==4.12.2

# Development Tools
python-dotenv==1.0.0
pytest==8.0.2